+++ examples/offscreen-common/src/commonAssets/fragment_shader.glsl
@@ -1,22 +1,24 @@
-#version 100
+#version 310 es
 
 precision mediump float;
 
 uniform sampler2D uTexture;
 uniform sampler2D uNormalMap;
 
-varying vec3 vCameraPosTan;
-varying vec3 vPosTan;
-varying vec2 vTexCoords;
+in vec3 vCameraPosTan;
+in vec3 vPosTan;
+in vec2 vTexCoords;
+
+out vec4 outColor;
 
 void main() {
     vec3 cameraDirection = normalize(vCameraPosTan - vPosTan);
 
-    vec4 color = texture2D(uTexture, vTexCoords);
-    vec3 normal = texture2D(uNormalMap, vTexCoords).rgb;
+    vec4 color = texture(uTexture, vTexCoords);
+    vec3 normal = texture(uNormalMap, vTexCoords).rgb;
     normal = normalize(normal * 2.0 - 1.0);
 
     float exposure = max(dot(cameraDirection, normal), 0.0) * 0.6 + 0.4;
 
-    gl_FragColor = vec4(color.rgb * exposure, 1.0);
+    outColor = vec4(color.rgb * exposure, 1.0);
 }